. objbase.cpp, defines.h, AppManagercli.cpp, rtappman.cpp: moved GetVmSize, GetCpuUsage and GetOpenFileCount to objbase, so that they can be shared with skeleton code in Python and Java.
